requirements.txt
pillow
torch>=0.4.0
numpy
torchvision
